Add color module

+++ src/main/java/bca/Hapepedia/entity/Color.java
@@ -13,23 +13,23 @@ import javax.persistence.Table;
 public class Color {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     @Column(length = 20, nullable = false, unique = true)
     private String name;
 
     public Color() {
     }
 
-    public Color(int id, String name) {
+    public Color(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
